% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsdelta.R
\name{epsdelta}
\alias{epsdelta}
\title{(epsilon, delta)-differential privacy trade off function}
\usage{
epsdelta(epsilon, delta = 0)
}
\arguments{
\item{epsilon}{The \eqn{\varepsilon} privacy parameter.}

\item{delta}{The \eqn{\delta} privacy parameter.}
}
\value{
An
}
\description{
Creates an \eqn{(\varepsilon, \delta)}-differential privacy trade off function
}
\details{
TODO description
}
\examples{
#gdp(0.5)
}
\references{
Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
