% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_epsdelta.R
\name{est_epsdelta}
\alias{est_epsdelta}
\title{Estimate (epsilon, delta)-differential privacy of trade off points}
\usage{
est_epsdelta(x, epsilon = NULL, delta = NULL, dp = 2L)
}
\arguments{
\item{x}{Add description from fdp to here.}

\item{epsilon}{If specified, is the fixed epsilon to use with bounding attempted by varying delta.}

\item{delta}{If specified, is the fixed delta to use with bounding attempted by varying epsilon.}

\item{dp}{The number of decimal places of accuracy to use when estimating the parameters of (epsilon, delta)-differential privacy.}
}
\value{
An
}
\description{
Estimates the (epsilon, delta)-differential privacy of a set of trade off points
}
\details{
TODO description
}
\examples{
#gdp(0.5)
}
\references{
Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
