% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gdp.R
\name{est_gdp}
\alias{est_gdp}
\title{Estimate Gaussian differential privacy of trade off points}
\usage{
est_gdp(x, dp = 2L)
}
\arguments{
\item{x}{The GDP privacy parameter.}

\item{dp}{The number of decimal places to which lower bounding Gaussian differential privacy function should be computed.}
}
\value{
A GDP function
}
\description{
Estimates the Gaussian differential privacy of a set of trade off points
}
\details{
This function finds the smallest GDP parameter \eqn{\mu} that lower bounds a given set trade-off points.
It uses numerical optimisation to identify the smallest \eqn{\mu} value such that the GDP trade-off function lower bounds all provided trade-off points.
The precision of the lower bound is controlled by the \code{dp} parameter, which specifies the number of decimal places for rounding.
}
\examples{
#gdp(0.5)
}
\references{
Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
