% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lap.R
\name{lap}
\alias{lap}
\title{Laplace differential privacy trade off function}
\usage{
lap(mu = 1)
}
\arguments{
\item{mu}{The Laplace privacy parameter.}
}
\value{
A Laplace trade-off function
}
\description{
Creates a Laplace differential privacy trade off function
}
\details{
TODO description
}
\examples{
#lap(1)
}
\references{
Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
