% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_gdp.R
\name{est_gdp}
\alias{est_gdp}
\title{Estimate Gaussian differential privacy of trade off points}
\usage{
est_gdp(x, dp = 2L)
}
\arguments{
\item{mu}{The GDP privacy parameter.}
}
\value{
A GDP function
}
\description{
Estimates the Gaussian differential privacy of a set of trade off points
}
\details{
TODO description
}
\examples{
#gdp(0.5)
}
\references{
Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
