% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp-package.R
\docType{package}
\name{fdp-package}
\alias{fdp-package}
\title{fdp: f-Differential Privacy and Gaussian Differential Privacy}
\description{
Constructs and visualises trade-off functions for f-differential privacy (f-DP) as introduced by Dong et al. (2022) \doi{10.1111/rssb.12454}. Supports Gaussian differential privacy, the f-DP generalisation of (epsilon, delta)-differential privacy, and accepts user-specified optimal type I / type II errors from which the lower convex hull trade-off function is automatically constructed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fdp.louisaslett.com/}
  \item Report bugs at \url{https://github.com/louisaslett/fdp/issues}
}

}
\author{
\strong{Maintainer}: Louis Aslett \email{louis.aslett@durham.ac.uk} (\href{https://orcid.org/0000-0003-2211-233X}{ORCID}) [copyright holder]

}
\keyword{internal}
