% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp.R
\name{fdp}
\alias{fdp}
\title{Plot f-differential privacy}
\usage{
fdp(...)
}
\arguments{
\item{...}{f-differential privacy inputs to plot.}
}
\value{
A ggplot2 object which will plot
}
\description{
Plots an f-differential privacy trade off function
}
\details{
TODO description
}
\examples{
#fdp(epsdelta(1,0.1), gdp(1), "a"=gdp(0.5), tst=data.frame(alpha=c(1,0.5,0),beta=c(0,0.3,1)), asd=data.frame(alpha=c(1,0.4,0),beta=c(0,0.51,1)))
}
\references{
Andrew, A. M. (1979). “Another efficient algorithm for convex hulls in two dimensions”. \emph{Information Processing Letters}, \strong{9}(5), 216–219. \doi{10.1016/0020-0190(79)90072-3}.

Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
