% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{gdp}
\alias{gdp}
\title{Gaussian differential privacy trade off function}
\usage{
gdp(mu = 1)
}
\arguments{
\item{mu}{The GDP privacy parameter.}
}
\value{
A GDP object
}
\description{
Creates a Gaussian differential privacy trade off function
}
\details{
TODO description
}
\examples{
#gdp(0.5)
}
\references{
Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
