% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_to_epsdelta.R
\name{gdp_to_epsdelta}
\alias{gdp_to_epsdelta}
\title{Exact (epsilon, delta)-differential privacy of GDP}
\usage{
gdp_to_epsdelta(mu = 0.5, epsilon = 1, dp = NULL)
}
\arguments{
\item{mu}{The parameter of GDP to match}

\item{epsilon}{The fixed target epsilon to use, with bounding computed by finding delta.}

\item{dp}{The number of decimal places of accuracy to use when estimating the delta parameter.}
}
\value{
An
}
\description{
Calculated exact (epsilon, delta)-differential privacy of a GDP using Corollary 1, p.16, in Dong et al. (2022).
}
\details{
TODO description
}
\examples{
#gdp(0.5)
}
\references{
Dong, J., Roth, A. and Su, W.J. (2022). “Gaussian Differential Privacy”. \emph{Journal of the Royal Statistical Society Series B}, \strong{84}(1), 3–37. \doi{10.1111/rssb.12454}.
}
